+++ server/auth.ts
@@ -39,7 +39,7 @@ import { isLocalNetworkOrigin } from './lib/local-origin'
 import { hashPassword, verifyPassword as verifyPasswordHash } from './lib/password'
 import { createDbProxy, createPlatformProxy } from './platform/context'
 import type { Database, Platform } from './platform/interface'
-import { loadCaptchaConfig } from './usecases/captcha'
+import { loadCaptchaConfig } from './usecases/site/captcha'
 import { checkTeamLimit, getEffectiveSignupMode } from './usecases/site/licensing'
 
 // better-auth's default password hasher is pure-JS scrypt from @noble/hashes,
