+++ server/usecases/site/system.ts
@@ -19,8 +19,8 @@ import { readCaptchaConfig } from '../../domain/captcha'
 import { hasFeature } from '../../domain/licensing'
 import { originFromRequestUrl } from '../../domain/site-public-origin'
 import { getAppVersion } from '../../version'
-import { loadCaptchaOptionValues } from '../captcha'
 import type { ActivityRepo, ChangelogProvider, InstanceRepo, LicenseBindingRepo, SystemOptionsRepo } from '../ports'
+import { loadCaptchaOptionValues } from './captcha'
 import { buildInstanceInfo, type runtimeInfo } from './instance-info'
 import { loadBindingState } from './licensing'
 import { getSitePublicOrigin } from './public-origin'
